+++ examples/lib/main.dart
@@ -12,7 +12,7 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -28,8 +28,8 @@ class MyApp extends StatelessWidget {
 
 class HomePage extends StatelessWidget {
   const HomePage({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -80,11 +80,11 @@ class HomePage extends StatelessWidget {
 
 class MenuEntry extends StatelessWidget {
   const MenuEntry({
-    Key? key,
+    super.key,
     required this.title,
     required this.imageName,
     required this.destination,
-  }) : super(key: key);
+  });
 
   final String title;
   final Widget destination;
@@ -122,7 +122,7 @@ class MenuEntry extends StatelessWidget {
                       textAlign: TextAlign.center,
                       style: Theme.of(context)
                           .textTheme
-                          .headline6!
+                          .titleLarge!
                           .copyWith(color: Colors.white),
                     ),
                   ),