+++ examples/lib/pages/masonry.dart
@@ -7,8 +7,8 @@ import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
 
 class MasonryPage extends StatefulWidget {
   const MasonryPage({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   State<MasonryPage> createState() => _MasonryPageState();